Add files via upload
Uploaded python_challenge homework for week 3.  There were two assignments.  The first assignment is contained in pybank.py.  This assignment takes the contents of a csv file and completes some simple financial analysis on the data.  The python script extracts the total number of months of data, Total values, Average change of balances month to month, and differences in profits on a rolling monthly basis.

The second assignment is contained in pypoll.py.   This assignment reads in voting data.  The script calculates election results:  total votes, votes per candidate, and displays the winner of the election.